Allow for termux/stable & termux/debug & ..
The script looks for .deb files in subfolders and takes the subfolder name as
COMPONENT.
.deb files located in the input_path directory (and not in a subfolder) are
still put into the "extras" component folder.

When updating the Release file, all component directories are taken into
account, so it's possible to only update for example termux/debug and keeping
termux/stable as is.